Move files to shared folders
